fix - can add items to cart after deleting them once
Earlier once items were deleted from cart, they couldn't be added again.

This was happening because the id was being passed on from the front end and the id associated with each product was the same.

In database tables once and id (PK) is used, then it can't be used again even after deletion.

changed the id to productId, now productId is used to identify the product and the id for the database entry is generated by the backend and stored in the id key of the object.